+++ server/tests/test_pipeline.py
@@ -12,6 +12,7 @@ from weaverbird.pipeline.pipeline import (
     Pipeline,
     PipelineStep,
     PipelineStepWithVariables,
+    PipelineWithRefs,
     PipelineWithVariables,
     remove_void_conditions_from_filter_steps,
     remove_void_conditions_from_mongo_steps,
@@ -598,3 +599,33 @@ def test_remove_void_conditions_from_filter_steps_with_combinations(
     expected_steps: list[PipelineStep | PipelineStepWithVariables],
 ) -> None:
     assert remove_void_conditions_from_filter_steps(steps) == expected_steps
+
+
+def test_pipeline_with_refs_variables_and_date_validity():
+    PipelineWithRefs(
+        steps=[
+            {"domain": {"type": "ref", "uid": "83ff1fa2-d186-4a7b-a53b-47c901a076c7"}, "name": "domain"},
+            {
+                "name": "append",
+                "pipelines": [
+                    {"type": "ref", "uid": "3ecb05aa-1312-4797-bf9a-65f736bb2993"},
+                    [
+                        {"name": "domain", "domain": "foo"},
+                        {
+                            "name": "filter",
+                            "condition": {
+                                "column": "date",
+                                "operator": "from",
+                                "value": {
+                                    "date": "{{TODAY}}",
+                                    "duration": "year",
+                                    "operator": "until",
+                                    "quantity": 1,
+                                },
+                            },
+                        },
+                    ],
+                ],
+            },
+        ]
+    )